shifumi/requirement.txt
black==20.8b1
PyYAML==5.3.1
